rename: Cfn::Status to CfnStatus
