--- lib/cfn/status/version.rb
@@ -1,5 +1,0 @@
-module Cfn
-  class Status
-    VERSION = "0.2.0"
-  end
-end
