+++ lib/cfn_status/version.rb
@@ -1,0 +1,3 @@
+class CfnStatus
+  VERSION = "0.2.0"
+end
